+++ programm/index.html
@@ -9,6 +9,7 @@
     <script src="javascript/script.js" defer type="module"></script>
     <script src="javascript/classes.js" defer type="module"></script>
     <script src="javascript/attacks.js" defer type="module"></script>
+    <script src="javascript/battlescene.js" defer type="module"></script>
     <script src="data/battlezones/map 1/battlezone 1.js" defer type="module"></script>
     <script src="data/battlezones/map 1/battlezone 2.js" defer type="module"></script>
     <script src="data/battlezones/map 1/battlezone 3.js" defer type="module"></script>
